requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
python-dotenv
transformers
torch
faiss-cpu